renamed `Recipe` module to `RecipeDetails`.
