+++ test/config/routes/app_named_routes_test.dart
@@ -1,1 +1,123 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:groceries/config/routes/app_named_routes.dart';
 
+void main() {
+  group(
+    'AppNamedRoutes',
+    () {
+      test(
+        'AppNamedRoutes have correct route names.',
+        () {
+          expect(
+            AppNamedRoutes.wrapper,
+            equals('wrapper'),
+          );
+          expect(
+            AppNamedRoutes.onboarding,
+            equals('onboarding'),
+          );
+          expect(
+            AppNamedRoutes.welcome,
+            equals('chooseAuth'),
+          );
+          expect(
+            AppNamedRoutes.login,
+            equals('login'),
+          );
+          expect(
+            AppNamedRoutes.register,
+            equals('register'),
+          );
+          expect(
+            AppNamedRoutes.forgotPassword,
+            equals('forgotPassword'),
+          );
+          expect(
+            AppNamedRoutes.resetPasswordSentSuccessfully,
+            equals('resetPasswordSentSuccessfully'),
+          );
+          expect(
+            AppNamedRoutes.root,
+            equals('root'),
+          );
+          expect(
+            AppNamedRoutes.home,
+            equals('home'),
+          );
+          expect(
+            AppNamedRoutes.recipes,
+            equals('recipes'),
+          );
+          expect(
+            AppNamedRoutes.recipeDetails,
+            equals('recipeDetails'),
+          );
+          expect(
+            AppNamedRoutes.createRecipe,
+            equals('createRecipe'),
+          );
+          expect(
+            AppNamedRoutes.addItems,
+            equals('addItems'),
+          );
+          expect(
+            AppNamedRoutes.notifications,
+            equals('notifications'),
+          );
+          expect(
+            AppNamedRoutes.groceryListDetails,
+            equals('groceryListDetails'),
+          );
+          expect(
+            AppNamedRoutes.goceryListSettings,
+            equals('goceryListSettings'),
+          );
+          expect(
+            AppNamedRoutes.createList,
+            equals('createList'),
+          );
+          expect(
+            AppNamedRoutes.inviteUsersToList,
+            equals('inviteUsersToList'),
+          );
+          expect(
+            AppNamedRoutes.listCreatedSuccessfully,
+            equals('listCreatedSuccessfully'),
+          );
+          expect(
+            AppNamedRoutes.listCreatedUnsuccessfully,
+            equals('listCreatedUnsuccessfully'),
+          );
+          expect(
+            AppNamedRoutes.settings,
+            equals('settings'),
+          );
+          expect(
+            AppNamedRoutes.accountSettings,
+            equals('accountSettings'),
+          );
+          expect(
+            AppNamedRoutes.deleteAccount,
+            equals('deleteAccount'),
+          );
+          expect(
+            AppNamedRoutes.appearanceSettings,
+            equals('appearanceSettings'),
+          );
+          expect(
+            AppNamedRoutes.notificationsSettings,
+            equals('notificationsSettings'),
+          );
+          expect(
+            AppNamedRoutes.additionalResources,
+            equals('additionalResources'),
+          );
+          expect(
+            AppNamedRoutes.pageNotFound,
+            equals('pageNotFound'),
+          );
+        },
+      );
+    },
+  );
+}
